requirements.txt
gradio==4.38.1
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.19.0 